combine weekplan and week view
